[backend] Fix user CSV domain
