+++ opencti-platform/opencti-graphql/src/modules/user/user-domain.ts
@@ -1,3 +1,4 @@
+import { v4 as uuid } from 'uuid';
 import type { AuthContext, AuthUser } from '../../types/user';
 import { addUser, findAll as findAllUser } from '../../domain/user';
 import { SYSTEM_USER } from '../../utils/access';
@@ -5,6 +6,9 @@ import type { BasicGroupEntity, BasicStoreCommon } from '../../types/store';
 import { findDefaultIngestionGroups } from '../../domain/group';
 import { FunctionalError, ValidationError } from '../../config/errors';
 import type { UserAddInput } from '../../generated/graphql';
+import { getEntityFromCache } from '../../database/cache';
+import type { BasicStoreSettings } from '../../types/settings';
+import { ENTITY_TYPE_SETTINGS } from '../../schema/internalObject';
 
 export const userAlreadyExists = async (context: AuthContext, name: string) => {
   // We use SYSTEM_USER because manage ingestion should be enough to create an ingestion Feed
@@ -24,24 +28,28 @@ export const userAlreadyExists = async (context: AuthContext, name: string) => {
   return users.length > 0;
 };
 
-export const createOnTheFlyUser = async (context: AuthContext, user: AuthUser, input: { userName: string, confidenceLevel: string | null | undefined }) => {
+export const createOnTheFlyUser = async (context: AuthContext, user: AuthUser, input: { userName: string, confidenceLevel: number | null | undefined }) => {
   const defaultIngestionGroups: BasicGroupEntity[] = await findDefaultIngestionGroups(context, user) as BasicGroupEntity[];
   if (defaultIngestionGroups.length < 1) {
     throw FunctionalError('You have not defined a default group for ingestion users', {});
   }
   const isUserAlreadyExisting = await userAlreadyExists(context, input.userName);
   if (isUserAlreadyExisting) {
-    throw FunctionalError('This service account already exists. Change the instance name to change the automatically created service account name', {});
+    throw FunctionalError('This user already exists. Change the feed\'s name to change the automatically created user\'s name', {});
   }
-  let userInput: UserAddInput;
-  userInput = {
+  const { platform_organization } = await getEntityFromCache<BasicStoreSettings>(context, SYSTEM_USER, ENTITY_TYPE_SETTINGS);
+
+  let userInput: UserAddInput = {
+    password: uuid(),
+    user_email: `automatic+${uuid()}@opencti.invalid`,
     name: input.userName,
     prevent_default_groups: true,
     groups: [defaultIngestionGroups[0].id],
-    user_service_account: true,
+    objectOrganization: platform_organization ? [platform_organization] : []
   };
+
   if (input.confidenceLevel) {
-    const userConfidence = parseFloat(input.confidenceLevel);
+    const userConfidence = input.confidenceLevel;
     if (userConfidence < 0 || userConfidence > 100 || !Number.isInteger(userConfidence)) {
       throw ValidationError('The confidence_level should be an integer between 0 and 100', 'confidence_level');
     }
